requirements.txt
django
psycopg2
Pillow
djangorestframework
django-cors-headers
